style: rename Exec to ExecCard
