+++ src/routes/people/ExecCard.svelte
